Copy STDOUT to reports instead of capturing it
It's quite surprising that, by default, this library prevents test
output from showing up in the normal place. At the same time, it is
nice that it shows up in the generated reports. We can get the best of
both worlds by setting stdout to an IO object that forwards on messages
to both the CI::Reporter StringIO, and the previous value of stdout.